+++ src/ResourceManager/RecoveryServices.Backup/Commands.RecoveryServices.Backup.Helpers/Conversions/ConversionHelpers.cs
@@ -14,11 +14,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models;
-using CmdletModels = Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models;
 using Microsoft.Azure.Commands.RecoveryServices.Backup.Properties;
 using ServiceClientModel = Microsoft.Azure.Management.RecoveryServices.Backup.Models;
 
@@ -41,16 +37,16 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
             if (protectionContainer != null &&
                 protectionContainer.Properties != null)
             {
-                if (protectionContainer.Properties.GetType().IsSubclassOf(typeof(ServiceClientModel.AzureIaaSVMProtectionContainer)))
+                if (protectionContainer.Properties.GetType().IsSubclassOf(typeof(ServiceClientModel.IaaSVMContainer)))
                 {
                     containerModel = new AzureVmContainer(protectionContainer);
                 }
-                if (protectionContainer.Properties.GetType() == typeof(ServiceClientModel.MabProtectionContainer))
+                if (protectionContainer.Properties.GetType() == typeof(ServiceClientModel.MabContainer))
                 {
                     containerModel = new MabContainer(protectionContainer);
                 }
                 else if (protectionContainer.Properties.GetType() ==
-                    typeof(ServiceClientModel.AzureSqlProtectionContainer))
+                    typeof(ServiceClientModel.AzureSqlContainer))
                 {
                     containerModel = new AzureSqlContainer(protectionContainer);
                 }
@@ -62,13 +58,22 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
         /// <summary>
         /// Helper function to convert ps backup engine model from service response.
         /// </summary>
-        public static BackupEngineBase GetBackupEngineModel(ServiceClientModel.BackupEngineResource backupEngine)
+        public static BackupEngineBase GetBackupEngineModel(
+            ServiceClientModel.BackupEngineBaseResource backupEngine)
         {
             BackupEngineBase backupEngineModel = null;
 
             if (backupEngine != null &&
                 backupEngine.Properties != null)
             {
+                string friendlyName = backupEngine.Properties.FriendlyName;
+                string backupManagementType =
+                    backupEngine.Properties.BackupManagementType.ToString();
+                string registrationStatus = backupEngine.Properties.RegistrationStatus;
+                string healthStatus = backupEngine.Properties.HealthStatus;
+                bool? canReRegister = backupEngine.Properties.CanReRegister;
+                string backupEngineId = backupEngine.Properties.BackupEngineId;
+
                 if (backupEngine.Properties.GetType() == (typeof(ServiceClientModel.DpmBackupEngine)))
                 {
                     backupEngineModel = new DpmBackupEngine(backupEngine);
@@ -100,7 +105,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
         /// <summary>
         /// Helper function to convert ps backup engine model list from service response.
         /// </summary>
-        public static List<BackupEngineBase> GetBackupEngineModelList(IEnumerable<ServiceClientModel.BackupEngineResource> backupEngines)
+        public static List<BackupEngineBase> GetBackupEngineModelList(
+            IEnumerable<ServiceClientModel.BackupEngineBaseResource> backupEngines)
         {
             List<BackupEngineBase> backupEngineModel = new List<BackupEngineBase>();
 
@@ -151,8 +157,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
 
                 policyModel = new AzureVmPolicy();
                 AzureVmPolicy iaasPolicyModel = policyModel as AzureVmPolicy;
-                iaasPolicyModel.WorkloadType = Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models.WorkloadType.AzureVM;
-                iaasPolicyModel.BackupManagementType = Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models.BackupManagementType.AzureVM;
+                iaasPolicyModel.WorkloadType = WorkloadType.AzureVM;
+                iaasPolicyModel.BackupManagementType = BackupManagementType.AzureVM;
                 iaasPolicyModel.RetentionPolicy = PolicyHelpers.GetPSLongTermRetentionPolicy((ServiceClientModel.LongTermRetentionPolicy)
                                                   ((ServiceClientModel.AzureIaaSVMProtectionPolicy)serviceClientResponse.Properties).RetentionPolicy);
                 iaasPolicyModel.SchedulePolicy = PolicyHelpers.GetPSSimpleSchedulePolicy((ServiceClientModel.SimpleSchedulePolicy)
@@ -175,8 +181,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
 
                 policyModel = new AzureSqlPolicy();
                 AzureSqlPolicy sqlPolicyModel = policyModel as AzureSqlPolicy;
-                sqlPolicyModel.WorkloadType = CmdletModels.WorkloadType.AzureSQLDatabase;
-                sqlPolicyModel.BackupManagementType = CmdletModels.BackupManagementType.AzureSQL;
+                sqlPolicyModel.WorkloadType = WorkloadType.AzureSQLDatabase;
+                sqlPolicyModel.BackupManagementType = BackupManagementType.AzureSQL;
 
                 ServiceClientModel.SimpleRetentionPolicy azureSqlRetentionPolicy =
                     (ServiceClientModel.SimpleRetentionPolicy)azureSqlPolicy.RetentionPolicy;
@@ -203,10 +209,9 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
         /// Helper function to convert ps backup policy list model from service response.
         /// </summary>
         public static List<PolicyBase> GetPolicyModelList(
-            ServiceClientModel.ProtectionPolicyListResponse serviceClientListResponse)
+            List<ServiceClientModel.ProtectionPolicyResource> serviceClientListResponse)
         {
-            if (serviceClientListResponse == null || serviceClientListResponse.ItemList == null ||
-               serviceClientListResponse.ItemList.Value == null || serviceClientListResponse.ItemList.Value.Count == 0)
+            if (serviceClientListResponse == null && serviceClientListResponse.Count == 0)
             {
                 Logger.Instance.WriteDebug("Received empty list of policies from service");
                 return null;
@@ -215,7 +220,8 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
             List<PolicyBase> policyModels = new List<PolicyBase>();
             PolicyBase policyModel = null;
 
-            foreach (ServiceClientModel.ProtectionPolicyResource resource in serviceClientListResponse.ItemList.Value)
+            foreach (ServiceClientModel.ProtectionPolicyResource resource
+                in serviceClientListResponse)
             {
                 policyModel = GetPolicyModel(resource);
                 if (policyModel != null)
@@ -260,18 +266,18 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
                     itemModel = new AzureVmItem(
                         protectedItem,
                         IdUtils.GetNameFromUri(containerUri),
-                        Cmdlets.Models.ContainerType.AzureVM,
+                        ContainerType.AzureVM,
                         policyName);
                 }
 
-                if (protectedItem.Properties.GetType() == 
+                if (protectedItem.Properties.GetType() ==
                     typeof(ServiceClientModel.AzureSqlProtectedItem))
                 {
                     ServiceClientModel.AzureSqlProtectedItem azureSqlProtectedItem =
                         (ServiceClientModel.AzureSqlProtectedItem)protectedItem.Properties;
                     string policyName = null;
                     string policyId = azureSqlProtectedItem.PolicyId;
-                    if (!String.IsNullOrEmpty(policyId))
+                    if (!string.IsNullOrEmpty(policyId))
                     {
                         Dictionary<UriEnums, string> keyVauleDict =
                         HelperUtils.ParseUri(policyId);
@@ -285,7 +291,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
                     itemModel = new AzureSqlItem(
                         protectedItem,
                         IdUtils.GetNameFromUri(containerUri),
-                        Cmdlets.Models.ContainerType.AzureSQL,
+                        ContainerType.AzureSQL,
                         policyName);
                 }
             }